+++ app/lib/sync/controller.js
@@ -11,6 +11,7 @@ const clientConfig = require('../config.js');
 var startup = StartupCtrl(env, clientConfig);
 
 var syncTimeout;
+var syncStartup = false;
 
 module.exports = function(env, tray) {
   var syncWindow;
@@ -80,25 +81,36 @@ module.exports = function(env, tray) {
   }
 
   function start() {
+
+    //return if sync is already running or is starting up
+    if(syncWindow || syncStartup) {
+      return logger.info('Sync: not starting sync because it is already running', {category: 'sync'});
+    }
+
+    syncStartup = true;
+
     //return if no user is logged in
-    if(clientConfig.get('loggedin') === false) return logger.info('Sync: not starting sync because no user logged in');
+    if(clientConfig.get('loggedin') === false) {
+      return logger.info('not starting sync because no user logged in', {category: 'sync'});
+    }
 
     //return if no network available
-    if(clientConfig.get('onLineState') === false) return logger.info('Sync: not starting because no network available');
+    if(clientConfig.get('onLineState') === false) {
+      return logger.info('not starting because no network available', {category: 'sync'});
+    }
 
     //return if sync has been paused
-    if(syncPaused) return logger.info('Sync: not starting sync because it has been paused')
+    if(syncPaused) {
+      return logger.info('Sync: not starting sync because it has been paused', {category: 'sync'});
+    }
 
-    //return if sync is already running
-    if(syncWindow) return logger.info('Sync: not starting sync because it is already running');
-
-    logger.info('Sync: starting sync');
+    logger.info('starting sync', {category: 'sync'});
     tray.syncStarted();
 
     startPowerSaveBlocker();
 
     startup.preSyncCheck().then(result => {
-      logger.info('Sync: pre sync check successfull');
+      logger.info('pre sync check successfull', {category: 'sync'});
 
       syncWindow = new BrowserWindow({
           width: 1000,
@@ -110,6 +122,8 @@ module.exports = function(env, tray) {
           transparent: false,
           skipTaskbar: true
       });
+
+      syncStartup = false;
 
       syncWindow.loadURL(url.format({
           pathname: path.join(__dirname, 'index.html'),
@@ -135,8 +149,13 @@ module.exports = function(env, tray) {
         tray.syncEnded();
       });
     }).catch(err => {
-      logger.error('Sync: pre sync check failed', err);
+      logger.error('pre sync check failed', {
+        category: 'sync',
+        error: err
+      });
+
       tray.syncEnded();
+      syncStartup = false;
       end(true);
     });
   }
@@ -148,7 +167,7 @@ module.exports = function(env, tray) {
       //do not set timeout when sync has been paused
       syncTimeout = setTimeout(() => {
         start();
-      }, ((env.sync && env.sync.interval) || 30) * 1000);
+      }, ((env.sync && env.sync.interval) || 5) * 1000);
     }
 
     if(syncWindow) syncWindow.close();