+++ app/ui/startup/controller.js
@@ -26,7 +26,7 @@ var balloonAutoLauncher = new autoLaunch({
     isHidden: true
 });
 
-var logger = require('../../lib/logger.js');
+const logger = require('../../lib/logger.js');
 
 module.exports = function(env, clientConfig) {
   var startupWindow;
@@ -36,8 +36,14 @@ module.exports = function(env, clientConfig) {
 
   function enableAutoLaunch() {
     return new Promise(function(resolve, reject) {
-      if(env.enableAutoLaunch === false) return resolve();
+      if(env.enableAutoLaunch === false) {
+        logger.debug('autolaunch is disabled, skip enabling autolaunch', {
+          category: 'startup'
+        });
+        return resolve();
+      }
 
+      logger.debug('verify that autolaunch is enabled', {category: 'startup'});
       balloonAutoLauncher.isEnabled().then(function(isEnabled) {
         if(!isEnabled) {
           balloonAutoLauncher.enable().then(function(isEnabled) {
@@ -60,7 +66,6 @@ module.exports = function(env, clientConfig) {
 
   function checkConfig() {
     return Promise.all([
-      firstTimeStart(),
       makeSureBalloonDirExists(),
       makeSureContextMenuExists(),
       enableAutoLaunch(),
@@ -72,24 +77,19 @@ module.exports = function(env, clientConfig) {
     return makeSureBalloonDirExists();
   }
 
-  function firstTimeStart() {
-    if(!clientConfig.get('blnUrl') || !clientConfig.get('apiUrl') || !clientConfig.hadConfig()
-     || !clientConfig.isActiveInstance()) {
-      return firstTimeWizard();
-    } else {
-      return Promise.resolve();
-    }
-  }
-
   function makeSureBalloonDirExists() {
     return new Promise(function(resolve, reject) {
       var balloonDir = clientConfig.get('balloonDir');
+      logger.debug('verify that balloonDir exists', {category: 'startup'});
 
       //make sure balloonDir dir exists
       //TODO pixtron - can this be combined with config-manager?
       fs.exists(balloonDir, (exists) => {
         if(exists === false) {
-          logger.info('Startup: BalloonDir does not exist. Reseting db, last-cursor and creating BalloonDir');
+          logger.info('BalloonDir does not exist. Reseting db, last-cursor and creating BalloonDir', {
+            category: 'startup'
+          });
+
           //when balloonDir does not exist, cursor and db have to be reset too
           return Promise.all([
             configManager.resetCursorAndDb(),
@@ -106,10 +106,17 @@ module.exports = function(env, clientConfig) {
     return new Promise(function(resolve, reject) {
       fsUtility.createBalloonDir(clientConfig.get('balloonDir'), clientConfig.get('homeDir'), (err) => {
         if(err) {
-          logger.error('Startup:', {err});
+          logger.error('failed create ballon dir', {
+            category: 'startup',
+            error: err
+          });
+
           reject(err);
         } else {
-          logger.info('Startup: balloonDir created');
+          logger.info('balloonDir created', {
+            category: 'startup'
+          });
+
           resolve();
         }
       });
@@ -131,8 +138,7 @@ module.exports = function(env, clientConfig) {
   }
 
   function authenticate() {
-    return new Promise(function(resolve, reject) {
-      if(!clientConfig.get('blnUrl')
+    if(!clientConfig.get('blnUrl')
         ||
         !clientConfig.get('apiUrl')
         ||
@@ -141,10 +147,15 @@ module.exports = function(env, clientConfig) {
         !clientConfig.hadConfig()
         ||
         !clientConfig.isActiveInstance()
-      ) {
-        return resolve();
-      }
+    ) {
+      logger.debug('skip startup authentication, first time wizard needs to be executed first', {
+          category: 'startup'
+      });
+      return firstTimeWizard();
+    }
 
+    return new Promise(function(resolve, reject) {
+      logger.debug('verify authentication', {category: 'startup'});
       auth.login(askCredentials).then(() => {
         resolve();
       }).catch((error) => {
@@ -155,7 +166,10 @@ module.exports = function(env, clientConfig) {
 
   function askCredentials() {
     return new Promise(function(resolve, reject) {
-      if(!clientConfig.isActiveInstance() && clientConfig.get('onLineState') === true) {
+      logger.debug('ask user for authentication credentials', {category: 'startup'});
+
+      if(clientConfig.get('onLineState') === true) {
+        logger.debug('waiting for user action', {category: 'startup'});
         if(!startupWindow) startupWindow = createStartupWindow();
 
         startupWindow.webContents.executeJavaScript(`switchView('auth')`);
@@ -172,7 +186,11 @@ module.exports = function(env, clientConfig) {
 
         ipcMain.removeAllListeners('startup-basic-auth');
         ipcMain.on('startup-basic-auth', function(event, username, password) {
-          logger.info('requested basic authentication', {username});
+          logger.info('requested basic authentication', {
+            category: 'startup',
+            username: username
+          });
+
           startupWindow.removeListener('closed', windowClosedByUserHandler);
           auth.basicAuth(username, password).then(() => {
             if(!clientConfig.hadConfig()) {
@@ -191,7 +209,11 @@ module.exports = function(env, clientConfig) {
         ipcMain.removeAllListeners('auth-oidc-signin');
         ipcMain.on('auth-oidc-signin', function(event, idp) {
           var idpConfig = env.auth.oidc[idp];
-          logger.info('requested oidc signin', {idpConfig});
+          logger.info('requested oidc signin', {
+            category: 'startup',
+            idp: idpConfig
+          });
+
           startupWindow.removeListener('closed', windowClosedByUserHandler);
           auth.oidcAuth(idpConfig).then(() => {
             if(!clientConfig.hadConfig()) {
@@ -207,6 +229,9 @@ module.exports = function(env, clientConfig) {
           });
         });
       } else {
+        logger.error('can not ask for authentication credentials, there is an active instance ongoing', {
+            category: 'startup'
+        });
         resolve();
       }
     });
@@ -233,7 +258,7 @@ module.exports = function(env, clientConfig) {
   }
 
   function welcomeWizard() {
-    logger.info('Startup settings: open requested');
+    logger.info('Startup settings: open requested', {category: 'startup'});
 
     return new Promise(function(resolve, reject) {
       if(!startupWindow) startupWindow = createStartupWindow();
@@ -258,7 +283,8 @@ module.exports = function(env, clientConfig) {
 
       ipcMain.removeAllListeners('startup-change-dir');
       ipcMain.on('startup-change-dir', function(event) {
-        logger.info('Startup Settings: change balloon dir');
+        logger.info('change balloon data directory', {category: 'startup'});
+
         dialog.showOpenDialog({
           title: 'Select balloon directory',
           defaultPath: clientConfig.balloonDir,
@@ -274,16 +300,22 @@ module.exports = function(env, clientConfig) {
 
       ipcMain.removeAllListeners('startup-selective-sync');
       ipcMain.on('startup-selective-sync', function(event) {
-        logger.info('Startup Settings: selective sync');
+        logger.info('open selective sync window', {category: 'startup'});
         openSelectiveSync();
       });
     });
   }
 
   function firstTimeWizard() {
-    logger.info('Startup settings: open requested');
+    logger.info('start first time wizard', {category: 'startup'});
 
     return new Promise(function(resolve, reject) {
+
+      if(clientConfig.isActiveInstance()) {
+        logger.info('first time wizard unlinked active instance', {category: 'startup'});
+        auth.logout();
+      }
+
       if(!startupWindow) startupWindow = createStartupWindow();
 
       startupWindow.webContents.executeJavaScript(`switchView('server')`);
@@ -302,13 +334,22 @@ module.exports = function(env, clientConfig) {
         clientConfig.set('onLineState', true);
 
         if(!env.blnUrl) {
-          logger.info('Startup Settings: change blnUrl', {blnUrl});
+          logger.info('change url to server', {
+            category: 'startup',
+            url: blnUrl
+          });
+
           clientConfig.setBlnUrl(blnUrl);
         }
 
         askCredentials().then(() => {
           resolve();
         }).catch((error) => {
+          logger.error('failed ask for credentials', {
+            category: 'startup',
+            error: error
+          });
+
           reject(error);
         });
 
@@ -318,7 +359,7 @@ module.exports = function(env, clientConfig) {
   }
 
   function openSelectiveSync() {
-    logger.info('Startup settings: open selective sync requested');
+    logger.info('Startup settings: open selective sync requested', {category: 'startup'});
 
     return new Promise(function(resolve, reject) {
       if(!selectiveWindow) selectiveWindow = createSelectiveWindow();
@@ -331,7 +372,10 @@ module.exports = function(env, clientConfig) {
 
       ipcMain.removeAllListeners('selective-apply');
       ipcMain.on('selective-apply', function(event, ids) {
-        logger.info('Startup Settings: apply selective sync', {ids});
+        logger.info('apply selective sync settings', {
+          category: 'startup',
+          data: ids
+        });
 
         clientConfig.ignoreNode(ids);
 
@@ -340,7 +384,7 @@ module.exports = function(env, clientConfig) {
 
       ipcMain.removeAllListeners('selective-cancel');
       ipcMain.on('selective-cancel', function(event) {
-        logger.info('Startup Settings: cancel selective sync');
+        logger.info('cancel selective sync settings', {category: 'startup'});
         selectiveWindow.close();
       });
     });
@@ -410,13 +454,13 @@ module.exports = function(env, clientConfig) {
 
       windowStates.closed('startup-settings');
 
-      logger.info('Startup settings: closed');
+      logger.debug('startup window closed', {category: 'startup'});
     });
 
     startupWindow.on('show', (event) => {
       windowStates.opened('startup-settings');
 
-      logger.info('Startup settings: opened');
+      logger.debug('startup window opened', {category: 'startup'});
     });
 
     if(env.name === 'development') {
@@ -430,7 +474,6 @@ module.exports = function(env, clientConfig) {
     checkConfig,
     preSyncCheck,
     showBalloonDir,
-    showNodeSettingsWindow,
-    askCredentials
+    showNodeSettingsWindow
   }
 }